[arm64] kvm: arm64: Handle PtrAuth traps early
Message-id: <20200717123419.323545-7-drjones@redhat.com>
Patchwork-id: 322243
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH v4 06/12] KVM: arm64: Handle PtrAuth traps early
Bugzilla: 1855788
RH-Acked-by: Gavin Shan <gshan@redhat.com>
RH-Acked-by: Mark Salter <msalter@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Marc Zyngier <maz@kernel.org>

The current way we deal with PtrAuth is a bit heavy handed:

- We forcefully save the host's keys on each vcpu_load()
- Handling the PtrAuth trap forces us to go all the way back
  to the exit handling code to just set the HCR bits

Overall, this is pretty cumbersome. A better approach would be
to handle it the same way we deal with the FPSIMD registers:

- On vcpu_load() disable PtrAuth for the guest
- On first use, save the host's keys, enable PtrAuth in the
  guest

Crucially, this can happen as a fixup, which is done very early
on exit. We can then reenter the guest immediately without
leaving the hypervisor role.

Another thing is that it simplify the rest of the host handling:
exiting all the way to the host means that the only possible
outcome for this trap is to inject an UNDEF.

Reviewed-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
(cherry picked from commit 29eb5a3c57f7e06d803bb44a0ce2f9ed79f39cd9)
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>