[fs] jbd2: clear JBD2_ABORT flag before journal_reset to update log tail info when load journal
Message-id: <20200721113914.1967-17-lczerner@redhat.com>
Patchwork-id: 322295
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 fs PATCH 16/86] jbd2: clear JBD2_ABORT flag before journal_reset to update log tail info when load journal
Bugzilla: 1856760
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

From: Kai Li <li.kai4@h3c.com>

BZ 1856760
https://bugzilla.redhat.com/show_bug.cgi?id=1856760

BREW 30205319
https://brewweb.devel.redhat.com/taskinfo?taskID=30205319

commit a09decff5c32060639a685581c380f51b14e1fc2

If the journal is dirty when the filesystem is mounted, jbd2 will replay
the journal but the journal superblock will not be updated by
journal_reset() because JBD2_ABORT flag is still set (it was set in
journal_init_common()). This is problematic because when a new transaction
is then committed, it will be recorded in block 1 (journal->j_tail was set
to 1 in journal_reset()). If unclean shutdown happens again before the
journal superblock is updated, the new recorded transaction will not be
replayed during the next mount (because of stale sb->s_start and
sb->s_sequence values) which can lead to filesystem corruption.

Fixes: 85e0c4e89c1b ("jbd2: if the journal is aborted then don't allow update of the log tail")
Signed-off-by: Kai Li <li.kai4@h3c.com>
Link: https://lore.kernel.org/r/20200111022542.5008-1-li.kai4@h3c.com
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>