[fs] ext4: Reserve revoke credits for freed blocks
Message-id: <20200721113914.1967-83-lczerner@redhat.com>
Patchwork-id: 322361
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 fs PATCH 82/86] ext4: Reserve revoke credits for freed blocks
Bugzilla: 1856760
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Eric Sandeen <esandeen@redhat.com>

From: Jan Kara <jack@suse.cz>

BZ 1856760
https://bugzilla.redhat.com/show_bug.cgi?id=1856760

BREW 30205319
https://brewweb.devel.redhat.com/taskinfo?taskID=30205319

commit 83448bdfb59731c2f54784ed3f4a93ff95be6e7e

So far we have reserved only relatively high fixed amount of revoke
credits for each transaction. We over-reserved by large amount for most
cases but when freeing large directories or files with data journalling,
the fixed amount is not enough. In fact the worst case estimate is
inconveniently large (maximum extent size) for freeing of one extent.

We fix this by doing proper estimate of the amount of blocks that need
to be revoked when removing blocks from the inode due to truncate or
hole punching and otherwise reserve just a small amount of revoke
credits for each transaction to accommodate freeing of xattrs block or
so.

Signed-off-by: Jan Kara <jack@suse.cz>
Link: https://lore.kernel.org/r/20191105164437.32602-23-jack@suse.cz
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>