[vhost] vsock/virtio: fix multiple packet delivery to monitoring devices
Message-id: <20200730082953.21337-8-sgarzare@redhat.com>
Patchwork-id: 323097
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH 7/8] vsock/virtio: fix multiple packet delivery to monitoring devices
Bugzilla: 1861762
RH-Acked-by: Cathy Avery <cavery@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>

In virtio_transport.c, if the virtqueue is full, the transmitting
packet is queued up and it will be sent in the next iteration.
This causes the same packet to be delivered multiple times to
monitoring devices.

We want to continue to deliver packets to monitoring devices before
it is put in the virtqueue, to avoid that replies can appear in the
packet capture before the transmitted packet.

This patch fixes the issue, adding a new flag (tap_delivered) in
struct virtio_vsock_pkt, to check if the packet is already delivered
to monitoring devices.

In vhost/vsock.c, we are splitting packets, so we must set
'tap_delivered' to false when we queue up the same virtio_vsock_pkt
to handle the remaining bytes.

Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit a78d163978567adc2733465289293dad479d842a)
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>