[x86] x86_64: Add gap to int3 to allow for call emulation
Message-id: <20200731164552.27098-3-jolsa@redhat.com>
Patchwork-id: 323179
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.3 BZ1850831 2/7] x86_64: Add gap to int3 to allow for call emulation
Bugzilla: 1850831
RH-Acked-by: Vladis Dronov <vdronov@redhat.com>
RH-Acked-by: Chris von Recklinghausen <crecklin@redhat.com>
RH-Acked-by: Jesper Brouer <brouer@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Hildenbrand <david@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1850831
https://bugzilla.redhat.com/show_bug.cgi?id=1850831

upstream
========
commit 2700fefdb2d9751c416ad56897e27d41e409324a
Author: Josh Poimboeuf <jpoimboe@redhat.com>
Date:   Fri Nov 30 12:39:17 2018 -0600

description
===========
To allow an int3 handler to emulate a call instruction, it must be able to
push a return address onto the stack. Add a gap to the stack to allow the
int3 handler to push the return address and change the return from int3 to
jump straight to the emulated called function target.

Link: http://lkml.kernel.org/r/20181130183917.hxmti5josgq4clti@treble
Link: http://lkml.kernel.org/r/20190502162133.GX2623@hirez.programming.kicks-ass.net

[
  Note, this is needed to allow Live Kernel Patching to not miss calling a
  patched function when tracing is enabled. -- Steven Rostedt
]

Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>