[security] selinux: compute genfs symlink context in case of CephFS
Message-id: <20200812061821.168594-2-omosnace@redhat.com>
Patchwork-id: 323654
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 PATCH v3 2/2] selinux: compute genfs symlink context in case of CephFS
Bugzilla: 1865800
RH-Acked-by: Xiubo Li <xiubli@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1865800
Upstream Status: RHEL-only

Do not skip the genfs handling on symlinks in case of CephFS. This fix
is needed only in RHEL(-8) because of another RHEL-only tweak for
CephFS, which we did for backwards compatibility (see [1]). In upstream
policy, CephFS has SECURITY_FS_USE_XATTR behavior specified in the
policy, so there it is not affected.

We could probably remove the S_ISLNK() check entirely, but there would
be a certain risk of regressions for other filesystems using the GENFS
behavior, so we are turning it off conditionally based on a RHEL-only
sbsec flag which we only set for Ceph superblocks. In upstream, it can
be turned off entirely via a policy capability [2].

Note that starting with RHEL-9, we plan to fully inherit the upstream
behavior, so this patch doesn't need to be carried on beyond RHEL-8.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1823764
[2] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=7470d0d13fb680bb82b40f18831f7d4ee7a4bb62

Signed-off-by: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>