[powerpc] rhel: powerpc: kvm: Increase HDEC threshold to enter guest
Message-id: <20200817025106.87431-1-dgibson@redhat.com>
Patchwork-id: 323958
Patchwork-instance: patchwork
O-Subject: [RHEL-8.3 kernel virt PATCH] RHEL: powerpc: kvm: Increase HDEC threshold to enter guest
Bugzilla: 1733467
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Greg Kurz <gkurz@redhat.com>

From: David Gibson <david@gibson.dropbear.id.au>

RHEL-only fix: There are a number of questions we still want to answer
about this situation before wrapping it up upstream.  However, in the
meantime this patch gives at least an adequate fix for a nasty bug
downstream.  We expect this to be obsoleted by the upstream fix
eventually, but it's straightforward enough that it shouldn't cause
complicated conflicts.

Before entering a guest, we need to set the HDEC to pull us out again
when the guest's time is up.  This needs some care, though, because the
HDEC is edge triggered, which means that if it expires before entering the
guest, the interrupt will be lost, meaning we stay in the guest
indefinitely (in practice, until the the hard lockup detector pulls us out
with an NMI).

For the POWER9, independent threads mode specific path, we attempt to
prevent that, by testing time has already expired before setting the HDEC
in kvmhv_load_regs_and_go().  However, that doesn't account for the case
where the timer expires between that test and the actual guest entry.
Preliminary instrumentation suggests that can take as long as 1.5s under
certain load conditions, and simply checking the HDEC value we're going to
load is positive isn't enough to guarantee that leeway.

That test here is sometimes masked by a test in kvmhv_p9_guest_entry(), its
caller.  That checks that the remaining time is at 1s.  However as noted
above that doesn't appear to be sufficient in all circumstances even
from the point HDEC is set, let alone this earlier point.

Therefore, increase the threshold we check for in both locations to 4s
(2048 timebase ticks).  This is a pretty crude approach, but it addresses
a real problem where guest load can trigger a host hard lockup.

We're hoping to refine this in future by gathering more data on exactly
how long these paths can take, and possibly by moving the check closer to
the actual guest entry point to reduce the variance.  Getting the details
for that might take some time however.

NOTE: For reasons I haven't yet tracked down yet, I haven't actually
managed to reproduce this on current upstream.  I have reproduced it on
RHEL kernels without obvious differences in this area.  I'm still trying
to determine what the cause of that difference is, but I think it's worth
applying this change as a precaution in the interim.

Upstream: Not merged, still under investigation
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1733467
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=30769575
Testing: Ran for ~1 week and wasn't able to reproduce the problem,
         which usually appears in 2-10 hours
Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>