[arm64] kvm: arm64: Flush the instruction cache if not unmapping the VM on reboot
Message-id: <20200817160820.106067-2-drjones@redhat.com>
Patchwork-id: 323965
Patchwork-instance: patchwork
O-Subject: [RHEL8.3 virt PATCH 1/2] KVM: arm64: Flush the instruction cache if not unmapping the VM on reboot
Bugzilla: 1869297
RH-Acked-by: Gavin Shan <gshan@redhat.com>
RH-Acked-by: Auger Eric <eric.auger@redhat.com>

From: Marc Zyngier <maz@kernel.org>

On a system with FWB, we don't need to unmap Stage-2 on reboot,
as even if userspace takes this opportunity to repaint the whole
of memory, FWB ensures that the data side stays consistent even
if the guest uses non-cacheable mappings.

However, the I-side is not necessarily coherent with the D-side
if CTR_EL0.DIC is 0. In this case, invalidate the i-cache to
preserve coherency.

Reported-by: Alexandru Elisei <alexandru.elisei@arm.com>
Reviewed-by: Alexandru Elisei <alexandru.elisei@arm.com>
Fixes: 892713e97ca1 ("KVM: arm64: Sidestep stage2_unmap_vm() on vcpu reset when S2FWB is supported")
Signed-off-by: Marc Zyngier <maz@kernel.org>
(cherry picked from commit 7ae2f3db6167aa7184529fdacd1de72619baf93b)
Signed-off-by: Andrew Jones <drjones@redhat.com>

Conflicts:
  We don't have cpus_have_final_cap()
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>