[kernel] function_graph: Create function_graph_enter() to consolidate architecture code
Message-id: <20200904151727.1016728-2-jolsa@redhat.com>
Patchwork-id: 325599
Patchwork-instance: patchwork
O-Subject: [PATCHv2 RHEL8.3 BZ1857599 1/4] function_graph: Create function_graph_enter() to consolidate architecture code
Bugzilla: 1857599
RH-Acked-by: Jesper Brouer <brouer@redhat.com>
RH-Acked-by: Toke Høiland-Jørgensen <toke@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1857599
https://bugzilla.redhat.com/show_bug.cgi?id=1857599

upstream
========
commit 8114865ff82e200b383e46821c25cb0625b842b5
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date:   Sun Nov 18 17:10:15 2018 -0500

description
===========
Currently all the architectures do basically the same thing in preparing the
function graph tracer on entry to a function. This code can be pulled into a
generic location and then this will allow the function graph tracer to be
fixed, as well as extended.

Create a new function graph helper function_graph_enter() that will call the
hook function (ftrace_graph_entry) and the shadow stack operation
(ftrace_push_return_trace), and remove the need of the architecture code to
manage the shadow stack.

This is needed to prepare for a fix of a design bug on how the curr_ret_stack
is used.

Signed-off-by:Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>