[x86] x86/function_graph: Simplify with function_graph_enter()
Message-id: <20200904151727.1016728-3-jolsa@redhat.com>
Patchwork-id: 325600
Patchwork-instance: patchwork
O-Subject: [PATCHv2 RHEL8.3 BZ1857599 2/4] x86/function_graph: Simplify with function_graph_enter()
Bugzilla: 1857599
RH-Acked-by: Jesper Brouer <brouer@redhat.com>
RH-Acked-by: Toke Høiland-Jørgensen <toke@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1857599
https://bugzilla.redhat.com/show_bug.cgi?id=1857599

upstream
========
commit 07f7175b43827640d1e69c9eded89aa089a234b4
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date:   Sun Nov 18 17:14:10 2018 -0500

description
===========
The function_graph_enter() function does the work of calling the function
graph hook function and the management of the shadow stack, simplifying the
work done in the architecture dependent prepare_ftrace_return().

Have x86 use the new code, and remove the shadow stack management as well as
having to set up the trace structure.

This is needed to prepare for a fix of a design bug on how the curr_ret_stack
is used.

Signed-off-by:Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Frantisek Hrbata <fhrbata@redhat.com>