[powerpc] powerpc/pseries/iommu: Allow bigger 64bit window by removing default DMA window
Message-id: <20200918192218.12727.2409.sendpatchset@sbest.bos.csb>
Patchwork-id: 327268
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.4 BZ1725198 4/4] powerpc/pseries/iommu: Allow bigger 64bit window by removing default DMA window
Bugzilla: 1725198
RH-Acked-by: Gopal Tiwari <gtiwari@redhat.com>
RH-Acked-by: Vladis Dronov <vdronov@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1725198
Upstream:
commit 8c0d51592f6f0123953633d1ecf21e843fce0bfd

https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?h=next&id=8c0d51592f6f0123953633d1ecf21e843fce0bfd

Description:
On LoPAR "DMA Window Manipulation Calls", it's recommended to remove the
default DMA window for the device, before attempting to configure a DDW,
in order to make the maximum resources available for the next DDW to be
created.

This is a requirement for using DDW on devices in which hypervisor
allows only one DMA window.

If setting up a new DDW fails anywhere after the removal of this
default DMA window, it's needed to restore the default DMA window.
For this, an implementation of ibm,reset-pe-dma-windows rtas call is
needed:

Platforms supporting the DDW option starting with LoPAR level 2.7 implement
ibm,ddw-extensions. The first extension available (index 2) carries the
token for ibm,reset-pe-dma-windows rtas call, which is used to restore
the default DMA window for a device, if it has been deleted.

It does so by resetting the TCE table allocation for the PE to it's
boot time value, available in "ibm,dma-window" device tree node.

Signed-off-by: Leonardo Bras <leobras.c@gmail.com>
Tested-by: David Dai <zdai@linux.vnet.ibm.com>
Reviewed-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/20200805030455.123024-5-leobras.c@gmail.com

Signed-off-by: Steve Best <sbest@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>