[cpuidle] cpuidle: pseries: Set the latency-hint before entering CEDE
Message-id: <20200922141945.20100-2-didoming@redhat.com>
Patchwork-id: 327532
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH BZ1608791 1/4] cpuidle: pseries: Set the latency-hint before entering CEDE
Bugzilla: 1608791
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1608791
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=31453909

As per the PAPR, each H_CEDE call is associated with a latency-hint to
be passed in the VPA field "cede_latency_hint". The CEDE states that
we were implicitly entering so far is CEDE with latency-hint = 0.

This patch explicitly sets the latency hint corresponding to the CEDE
state that we are currently entering. While at it, we save the
previous hint, to be restored once we wakeup from CEDE. This will be
required in the future when we expose extended-cede states through the
cpuidle framework, where each of them will have a different
cede-latency hint.

commit 3af0ada7dd98c6da35c1fd7f107af3b9aa5e904c

Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
[mpe: Make cede_latency_hint static]
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/1596087177-30329-2-git-send-email-ego@linux.vnet.ibm.com
Signed-off-by: Diego Domingos <didoming@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>