[cpuidle] cpuidle: pseries: Fix CEDE latency conversion from tb to us
Message-id: <20200922141945.20100-5-didoming@redhat.com>
Patchwork-id: 327535
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH BZ1608791 4/4] cpuidle: pseries: Fix CEDE latency conversion from tb to us
Bugzilla: 1608791
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1608791
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=31453909

Commit d947fb4c965c ("cpuidle: pseries: Fixup exit latency for
CEDE(0)") sets the exit latency of CEDE(0) based on the latency values
of the Extended CEDE states advertised by the platform. The values
advertised by the platform are in timebase ticks. However the cpuidle
framework requires the latency values in microseconds.

If the tb-ticks value advertised by the platform correspond to a value
smaller than 1us, during the conversion from tb-ticks to microseconds,
in the current code, the result becomes zero. This is incorrect as it
puts a CEDE state on par with the snooze state.

This patch fixes this by rounding up the result obtained while
converting the latency value from tb-ticks to microseconds. It also
prints a warning in case we discover an extended-cede state with
wakeup latency to be 0. In such a case, ensure that CEDE(0) has a
non-zero wakeup latency.

commit 1d3ee7df009a46440c58508b8819213c09503acd

Fixes: d947fb4c965c ("cpuidle: pseries: Fixup exit latency for CEDE(0)")
Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Reviewed-by: Vaidyanathan Srinivasan <svaidy@linux.ibm.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/1599125247-28488-1-git-send-email-ego@linux.vnet.ibm.com
Signed-off-by: Diego Domingos <didoming@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>