[x86] efi/x86: Limit EFI old memory map to SGI UV machines
Message-id: <20201001185358.23889-206-crecklin@redhat.com>
Patchwork-id: 329525
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 bz 1754739 v7 patch 205/226] efi/x86: Limit EFI old memory map to SGI UV machines
Bugzilla: 1754739
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Waiman Long <longman@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Conflicts:  RHEL8 has an additional instalce of efi_enabled(EFI_OLD_MEMMAP).
	Change it to efi_have_uv1_memmap()

We carry a quirk in the x86 EFI code to switch back to an older
method of mapping the EFI runtime services memory regions, because
it was deemed risky at the time to implement a new method without
providing a fallback to the old method in case problems arose.

Such problems did arise, but they appear to be limited to SGI UV1
machines, and so these are the only ones for which the fallback gets
enabled automatically (via a DMI quirk). The fallback can be enabled
manually as well, by passing efi=old_map, but there is very little
evidence that suggests that this is something that is being relied
upon in the field.

Given that UV1 support is not enabled by default by the distros
(Ubuntu, Fedora), there is no point in carrying this fallback code
all the time if there are no other users. So let's move it into the
UV support code, and document that efi=old_map now requires this
support code to be enabled.

Note that efi=old_map has been used in the past on other SGI UV
machines to work around kernel regressions in production, so we
keep the option to enable it by hand, but only if the kernel was
built with UV support.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Ingo Molnar <mingo@kernel.org>
Link: https://lore.kernel.org/r/20200113172245.27925-8-ardb@kernel.org
(cherry picked from commit 1f299fad1e312947c974c6a1d8a3a484f27a6111)
Signed-off-by: Chris von Recklinghausen <crecklin@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>