[powerpc] powerpc/pseries: Add KVM guest doorbell restrictions
Message-id: <20201007130651.27720-4-didoming@redhat.com>
Patchwork-id: 330178
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH BZ1870221 3/3] powerpc/pseries: Add KVM guest doorbell restrictions
Bugzilla: 1870221
RH-Acked-by: David Gibson <dgibson@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1870221
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=31642010

KVM guests have certain restrictions and performance quirks when using
doorbells. This patch moves the EPAPR KVM guest test so it can be shared
with PSERIES, and uses that in doorbell setup code to apply the KVM
guest quirks and  improves IPI performance for two cases:

 - PowerVM guests may now use doorbells even if they are secure.

 - KVM guests no longer use doorbells if XIVE is available.

There is a valid complaint that "KVM guest" is not a very reasonable
thing to test for, it's preferable for the hypervisor to advertise
particular behaviours to the guest so they could change if the
hypervisor implementation or configuration changes. However in this case
we were already assuming a KVM guest worst case, so this patch is about
containing those quirks. If KVM later advertises fast doorbells, we
should test for that and override the quirks.

commit 107c55005fbd5243ee31fb13b6f166cde9e3ade1

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Tested-by: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/20200726035155.1424103-4-npiggin@gmail.com
Signed-off-by: Diego Domingos <didoming@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>