[scsi] scsi: libsas: Inject revalidate event for root port event
Message-id: <20201015152939.11968-7-emilne@redhat.com>
Patchwork-id: 331559
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 06/78] scsi: libsas: Inject revalidate event for root port event
Bugzilla: 1885733
RH-Acked-by: Ming Lei <ming.lei@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

According to the SAS spec, an expander device shall transmit BROADCAST
(CHANGE) from at least one phy in each expander port other than the
expander port that is the cause for transmitting BROADCAST (CHANGE).

As such, for when the link is lost for a root PHY attached to an expander
PHY, we get no broadcast event.

This causes an issue for libsas, in that we will not revalidate the domain
for these events.

As a solution, for when a root PHY is formed or deformed from a root port,
insert a broadcast event to trigger a domain revalidation.

Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 085f104a83d565097644889bf1f6f1aa6d345cb5)
Signed-off-by: Ewan D. Milne <emilne@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>