[scsi] scsi: scsi_debug: Disallow zone sizes that are not powers of 2
Message-id: <20201015152939.11968-43-emilne@redhat.com>
Patchwork-id: 331590
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 42/78] scsi: scsi_debug: Disallow zone sizes that are not powers of 2
Bugzilla: 1885733
RH-Acked-by: Ming Lei <ming.lei@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Allowing a non-power-of-2 zone size forces the use of direct division
operations of 64-bit sector values to obtain a zone number or number of
zones. Doing so without using do_div() leads to compilation errors on
32-bit architectures.

Devices with a zone size that is not a power of 2 do not exist today so
allowing their emulation is of limited interest as the sd driver will not
support them anyway. To fix this compilation error, instead of using
do_div() for sector values divisions, simply disallow zone size values that
are not a power of 2.

[mkp: commit desc]

Link: https://lore.kernel.org/r/20200507023526.221574-1-damien.lemoal@wdc.com
Fixes: 98e0a689868c ("scsi: scsi_debug: Add zone_size_mb module parameter")
Fixes: f0d1cf9378bd ("scsi: scsi_debug: Add ZBC zone commands")
Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Acked-by: Geert Uytterhoeven <geert@linux-m68k.org>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
(cherry picked from commit 108e36f0d8bf91839613d8053a6d1354965801b0)
Signed-off-by: Ewan D. Milne <emilne@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>