[sound] ALSA: memalloc: Make SG-buffer helper usable for continuous buffer, too
Message-id: <1602843533499326-490-patchset-jkysela@redhat.com>
Patchwork-id: 332121
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 0489/1534] ALSA: memalloc: Make SG-buffer helper usable for continuous buffer, too
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

We have a few helper functions for making the access to the buffer
address easier on SG-buffer.  Those are specific to the buffer that is
allocated with SG-buffer type, and it makes hard to use both SG and
non-SG buffers in the same code.

This patch adds a few simple checks and lets the helpers to deal with
both SG- and continuous buffers gracefully.  It's a preliminary step
for the upcoming patch that mimics the buffer type on the fly.

Link: https://lore.kernel.org/r/20200615160045.2703-4-tiwai@suse.de
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 2a1f3368bff609504cdc984cdb7cef467bb0b2b0)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>