[sound] ALSA: ASoC: samsung: Add driver for Aries boards
Message-id: <1602843533499326-514-patchset-jkysela@redhat.com>
Patchwork-id: 332131
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 0513/1534] ALSA: ASoC: samsung: Add driver for Aries boards
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Samsung Aries boards have a WM8994 codec connected to the Samsung
I2S controller, the BT codec, and the cellular modem.  Jack detection
is done by a combination of an ADC, GPIOs, and an extcon device for
the USB dock.  There is also a GPIO for selection between the Mic
path and the TV out path on the headphone jack.

There are two main variants, one with an FM radio and where the modem
is the master and one without a radio and the modem is the slave.

Signed-off-by: Jonathan Bakker <xc-racer2@live.ca>
Link: https://lore.kernel.org/r/BN6PR04MB06608CBF03EF27B70B175978A39F0@BN6PR04MB0660.namprd04.prod.outlook.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 7a3a7671fa6c7e90aff5f4242add2a40587b85ef)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>