[sound] ALSA: ASoC: rt5670: Fix dac- and adc- vol-tlv values being off by a factor of 10
Message-id: <1602843533499326-634-patchset-jkysela@redhat.com>
Patchwork-id: 332361
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 0633/1534] ALSA: ASoC: rt5670: Fix dac- and adc- vol-tlv values being off by a factor of 10
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

The adc_vol_tlv volume-control has a range from -17.625 dB to +30 dB,
not -176.25 dB to + 300 dB. This wrong scale is esp. a problem in userspace
apps which translate the dB scale to a linear scale. With the logarithmic
dB scale being of by a factor of 10 we loose all precision in the lower
area of the range when apps translate things to a linear scale.

E.g. the 0 dB default, which corresponds with a value of 47 of the
0 - 127 range for the control, would be shown as 0/100 in alsa-mixer.

Since the centi-dB values used in the TLV struct cannot represent the
0.375 dB step size used by these controls, change the TLV definition
for them to specify a min and max value instead of min + stepsize.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Link: https://lore.kernel.org/r/20200628155231.71089-5-hdegoede@redhat.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 3f31f7d9b5404a10648abe536c8b408bfb4502e1)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>