[sound] ALSA: ASoC: topology: use break on errors, not continue
Message-id: <1602843533499326-727-patchset-jkysela@redhat.com>
Patchwork-id: 332348
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 0726/1534] ALSA: ASoC: topology: use break on errors, not continue
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Since the beginning of the topology, the code continues to the next
object even when an error is detected.

The topology should be handled with an all-or-nothing design, loading
a partially valid topology is a sure way to get bug reports that are
difficult to deal with.

Changing the behavior may break previous solutions and expose problems
in topology files delivered in the past, so it's probably not wise to
add this patch to stable branches without revalidation.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Reviewed-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
Link: https://lore.kernel.org/r/20200707203749.113883-4-pierre-louis.bossart@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 129fc2ba01c8322575173cc97afa324e54a5d4ce)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>