[sound] ALSA: echoaudio: Remove redundant check
Message-id: <1602843533499326-738-patchset-jkysela@redhat.com>
Patchwork-id: 332383
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 0737/1534] ALSA: echoaudio: Remove redundant check
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

This check is always false, as it's not the responsibilty of the
device-specific code to make this check. It is already checked
in snd_echo_digital_mode_put.

I do not have a Mona interface to test this change.

This patch is in preparation for follow-up patch to modify the
behavior of "opencount".

Signed-off-by: Mark Hills <mark@xwax.org>
Link: https://lore.kernel.org/r/20200708101848.3457-1-mark@xwax.org
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit db091b0e14239f0ebd5709f3f9b5a1d662483adc)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>