[sound] ALSA: ASoC: fsl_spdif: Clear the validity bit for TX
Message-id: <1602843533499326-782-patchset-jkysela@redhat.com>
Patchwork-id: 332600
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 0781/1534] ALSA: ASoC: fsl_spdif: Clear the validity bit for TX
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

In IEC958 spec, "The validity bit is logical "0" if the
information in the main data field is reliable, and it
is logical "1" if it is not".

The default value of "ValCtrl" is zero, which means
"Outgoing Validity always set", then all the data is not
reliable, then some spdif sink device will drop the data.

So set "ValCtrl" to 1, that is to clear "Outgoing Validity"
in default.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Link: https://lore.kernel.org/r/1594112066-31297-2-git-send-email-shengjiu.wang@nxp.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 055b082156704b85a059770359d6cdedfb24831d)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>