[sound] ALSA: hda - fix the micmute led status for Lenovo ThinkCentre AIO
Message-id: <1602843533499326-1022-patchset-jkysela@redhat.com>
Patchwork-id: 332656
Patchwork-instance: patchwork
O-Subject: [RHEL 8.4 PATCH 1021/1534] ALSA: hda - fix the micmute led status for Lenovo ThinkCentre AIO
Bugzilla: 1869536
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

After installing the Ubuntu Linux, the micmute led status is not
correct. Users expect that the led is on if the capture is disabled,
but with the current kernel, the led is off with the capture disabled.

We tried the old linux kernel like linux-4.15, there is no this issue.
It looks like we introduced this issue when switching to the led_cdev.

Cc: <stable@vger.kernel.org>
Signed-off-by: Hui Wang <hui.wang@canonical.com>
Link: https://lore.kernel.org/r/20200810021659.7429-1-hui.wang@canonical.com
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit 386a6539992b82fe9ac4f9dc3f548956fd894d8c)
Bugzilla: 1869536
Signed-off-by: Jan Stancek <jstancek@redhat.com>